Fix last commit Add tests for HouseholdService

+++ src/main/java/com/tokyo/expensetracker/model/Role.java
@@ -9,6 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.AssertTrue;
 import javax.validation.constraints.NotEmpty;
+import java.util.Locale;
 
 @Entity
 @Getter
@@ -17,15 +18,15 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private byte id;
 
     @NotEmpty
     private String name;
 
     public Role() {}
 
-    public Role(Long id, String name) {
-        this.id = id;
+    public Role(int id, String name) {
+        this.id = (byte) id;
         this.name = name.toLowerCase();
     }
 
@@ -35,9 +36,8 @@ public class Role {
 
     @AssertTrue(message = "Must be any of the following values {'staff', 'co-Owner', 'owner'}")
     private boolean isValid(){
-         return name.equalsIgnoreCase("Staff")
-                 || name.equalsIgnoreCase("Co-Owner")
-                 || name.equalsIgnoreCase("Owner");
+        return name.equalsIgnoreCase("Staff")
+                || name.equalsIgnoreCase("Co-Owner")
+                || name.equalsIgnoreCase("Owner");
     }
 }
-
